requirements.txt
pytesseract==0.3.13
fastapi==0.115.7
pydantic==2.9.2
pillow==10.4.0
openai==1.58.1